moved delete confirmation to shared
