+++ src/app/admin_modules/admin_dashboard/dashboard/admin-dashboard/feature-flags/feature-flags.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { UserDataService } from '../../../services/user-data.service';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { DeleteConfirmModalComponent } from '../modal/delete-confirm-modal/delete-confirm-modal.component';
+import { DeleteConfirmModalComponent } from '../../../../../shared/modals/delete-confirm-modal/delete-confirm-modal.component';
 import { CreateOrUpdateFeatureFlagsComponent } from '../modal/create-or-update-feature-flags/create-or-update-feature-flags.component';
 import { UpdateJsonComponent } from '../modal/update-json/update-json.component';
 import { ViewJsonComponent } from '../modal/view-json/view-json.component';
