+++ src/app/shared/modals/delete-confirm-modal/delete-confirm-modal.component.scss
